refactor: consolidate article mappers and rename ArticleInDb to ArticleRow
