+++ src/articles/interfaces/article-row.interface.ts
@@ -5,8 +5,7 @@ import type { ArticleTag } from '@/tags/tags.schema';
 
 type ArticleFavoritedBy = InferSelectModel<typeof favoriteArticles>;
 
-export interface ArticleInDb
-  extends Omit<InferSelectModel<typeof articles>, 'authorId'> {
+export interface ArticleRow extends Omit<InferSelectModel<typeof articles>, 'authorId'> {
   author: Profile;
   favoritedBy: ArticleFavoritedBy[];
   tags: ArticleTag[];
